requirements.txt
flask==3.0.0
python-dotenv==1.0.1

openai==1.6.1

requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

simhash==2.1.2
